+++ sources/org/telegram/tgnet/TLRPC$TL_decryptedMessageMediaPhoto.java
@@ -25,7 +25,7 @@ public class TLRPC$TL_decryptedMessageMediaPhoto extends TLRPC$DecryptedMessageM
         abstractSerializedData.writeInt32(this.thumb_h);
         abstractSerializedData.writeInt32(this.w);
         abstractSerializedData.writeInt32(this.h);
-        abstractSerializedData.writeInt32(this.size);
+        abstractSerializedData.writeInt32((int) this.size);
         abstractSerializedData.writeByteArray(this.key);
         abstractSerializedData.writeByteArray(this.iv);
         abstractSerializedData.writeString(this.caption);
